requirements.txt
fastapi
pydantic
typing
openai-whisper